splitting registration and refactoring vocab.py module
