+++ torchtext/vocab/__init__.py
@@ -1,0 +1,23 @@
+from .vocab import Vocab
+
+from .vectors import(
+    GloVe,
+    FastText,
+    CharNGram,
+    pretrained_aliases,
+    Vectors,
+)
+
+from .vocab_factory import(
+    vocab,
+    build_vocab_from_iterator,
+)
+
+__all__ = ["Vocab",
+           "vocab",
+           "build_vocab_from_iterator",
+           "GloVe",
+           "FastText",
+           "CharNGram",
+           "pretrained_aliases",
+	   "Vectors"]
